pyproject.toml
[build-system]
requires = ["hatchling>=1.8.0"]
build-backend = "hatchling.build"

[project]
name = "deflow"
description = "Lightweight Declarative Data Workflow Framework"
readme = "README.md"
license = "MIT"
authors = [ { name = "ddeutils", email = "korawich.anu@gmail.com" } ]
keywords = [ 'framework', 'declarative', 'workflow' ]
classifiers = [
    "Topic :: Utilities",
    "Natural Language :: English",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.9.13"
dependencies = [
    "ddeutil-workflow==0.0.40",
    "python-dotenv==1.1.0",
]
dynamic = ["version"]

[project.scripts]
deflow = "deflow.__main__:main"

[tool.hatch.version]
scheme = "standard"
source = "code"
path = "deflow/__about__.py"

[tool.hatch.build.targets.sdist]
exclude = [
    "/.github",
    "/.pre-commit-config.yaml",
    "/docs",
    "/tests",
]

[tool.hatch.build.targets.wheel]
packages = ["deflow"]

[tool.shelf.version]
version = "./deflow/__about__.py"
changelog = "./docs/changelog.md"
commit_subject_format = "{emoji} {subject}"
commit_msg_format = "- {subject}"

[tool.coverage.run]
branch = true
relative_files = true
concurrency = [ "thread", "multiprocessing" ]
source = [ "deflow" ]
omit = [
    "deflow/__version__.py",
    "deflow/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "raise NotImplementedError",
    "pragma: no cov",
]

[tool.pytest.ini_options]
pythonpath = ["deflow"]
console_output_style = "count"
addopts = [
    "--strict-config",
    "--strict-markers",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::pytest.PytestUnraisableExceptionWarning",
]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)-7s] %(message)-120s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y%m%d %H:%M:%S"

[tool.black]
line-length = 80
target-version = [ 'py39' ]
exclude = """
(
    /(
        \\.git
        | \\.eggs
        | \\.__pycache__
        | \\.idea
        | \\.ruff_cache
        | \\.mypy_cache
        | \\.pytest_cache
        | \\.venv
        | build
        | dist
        | venv
    )/
)
"""

[project.urls]
Homepage = "https://github.com/ddeutils/deflow/"
"Source Code" = "https://github.com/ddeutils/deflow/"

[tool.ruff]
line-length = 80
exclude = [
    "__pypackages__",
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "build",
    "dist",
    "venv",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "F401",
    "F403",
]
